+++ mocks/releaser.go
@@ -21,12 +21,11 @@ type Releaser struct {
 	amendCommitReturnsOnCall map[int]struct {
 		result1 error
 	}
-	CommitAndReleaseStub        func(context.Context, git.VersionBump, string) error
+	CommitAndReleaseStub        func(context.Context, git.VersionBump) error
 	commitAndReleaseMutex       sync.RWMutex
 	commitAndReleaseArgsForCall []struct {
 		arg1 context.Context
 		arg2 git.VersionBump
-		arg3 string
 	}
 	commitAndReleaseReturns struct {
 		result1 error
@@ -162,20 +161,19 @@ func (fake *Releaser) AmendCommitReturnsOnCall(i int, result1 error) {
 	}{result1}
 }
 
-func (fake *Releaser) CommitAndRelease(arg1 context.Context, arg2 git.VersionBump, arg3 string) error {
+func (fake *Releaser) CommitAndRelease(arg1 context.Context, arg2 git.VersionBump) error {
 	fake.commitAndReleaseMutex.Lock()
 	ret, specificReturn := fake.commitAndReleaseReturnsOnCall[len(fake.commitAndReleaseArgsForCall)]
 	fake.commitAndReleaseArgsForCall = append(fake.commitAndReleaseArgsForCall, struct {
 		arg1 context.Context
 		arg2 git.VersionBump
-		arg3 string
-	}{arg1, arg2, arg3})
+	}{arg1, arg2})
 	stub := fake.CommitAndReleaseStub
 	fakeReturns := fake.commitAndReleaseReturns
-	fake.recordInvocation("CommitAndRelease", []interface{}{arg1, arg2, arg3})
+	fake.recordInvocation("CommitAndRelease", []interface{}{arg1, arg2})
 	fake.commitAndReleaseMutex.Unlock()
 	if stub != nil {
-		return stub(arg1, arg2, arg3)
+		return stub(arg1, arg2)
 	}
 	if specificReturn {
 		return ret.result1
@@ -189,17 +187,17 @@ func (fake *Releaser) CommitAndReleaseCallCount() int {
 	return len(fake.commitAndReleaseArgsForCall)
 }
 
-func (fake *Releaser) CommitAndReleaseCalls(stub func(context.Context, git.VersionBump, string) error) {
+func (fake *Releaser) CommitAndReleaseCalls(stub func(context.Context, git.VersionBump) error) {
 	fake.commitAndReleaseMutex.Lock()
 	defer fake.commitAndReleaseMutex.Unlock()
 	fake.CommitAndReleaseStub = stub
 }
 
-func (fake *Releaser) CommitAndReleaseArgsForCall(i int) (context.Context, git.VersionBump, string) {
+func (fake *Releaser) CommitAndReleaseArgsForCall(i int) (context.Context, git.VersionBump) {
 	fake.commitAndReleaseMutex.RLock()
 	defer fake.commitAndReleaseMutex.RUnlock()
 	argsForCall := fake.commitAndReleaseArgsForCall[i]
-	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3
+	return argsForCall.arg1, argsForCall.arg2
 }
 
 func (fake *Releaser) CommitAndReleaseReturns(result1 error) {