+++ pkg/processor/processor_test.go
@@ -480,7 +480,7 @@ var _ = Describe("Processor", func() {
 		}, 2*time.Second, 50*time.Millisecond).Should(Equal(1))
 
 		// Verify PatchBump was used
-		_, bump, _ := mockReleaser.CommitAndReleaseArgsForCall(0)
+		_, bump := mockReleaser.CommitAndReleaseArgsForCall(0)
 		Expect(bump).To(Equal(git.PatchBump))
 
 		// Verify CommitOnly was NOT called
@@ -568,7 +568,7 @@ var _ = Describe("Processor", func() {
 		}, 2*time.Second, 50*time.Millisecond).Should(Equal(1))
 
 		// Verify MinorBump was used
-		_, bump, _ := mockReleaser.CommitAndReleaseArgsForCall(0)
+		_, bump := mockReleaser.CommitAndReleaseArgsForCall(0)
 		Expect(bump).To(Equal(git.MinorBump))
 
 		cancel()
@@ -1052,7 +1052,8 @@ var _ = Describe("Processor", func() {
 			// Verify CommitOnly was called, not CommitAndRelease
 			Expect(mockReleaser.CommitOnlyCallCount()).To(Equal(1))
 			Expect(mockReleaser.CommitAndReleaseCallCount()).To(Equal(0))
-			Expect(mockReleaser.HasChangelogCallCount()).To(Equal(0))
+			// HasChangelog is called once for changelog suffix check during content assembly
+			Expect(mockReleaser.HasChangelogCallCount()).To(Equal(1))
 
 			cancel()
 		})
@@ -1480,7 +1481,8 @@ var _ = Describe("Processor", func() {
 			// Verify CommitOnly was called, not CommitAndRelease
 			Expect(mockReleaser.CommitOnlyCallCount()).To(Equal(1))
 			Expect(mockReleaser.CommitAndReleaseCallCount()).To(Equal(0))
-			Expect(mockReleaser.HasChangelogCallCount()).To(Equal(0))
+			// HasChangelog is called once for changelog suffix check during content assembly
+			Expect(mockReleaser.HasChangelogCallCount()).To(Equal(1))
 
 			cancel()
 		})